+++ src/main/java/com/example/EcoGo/repository/BadgeRepository.java
@@ -1,4 +1,5 @@
 package com.example.EcoGo.repository;
+
 import com.example.EcoGo.model.Badge;
 import org.springframework.data.mongodb.repository.MongoRepository;
 import org.springframework.stereotype.Repository;
@@ -9,8 +10,8 @@ import java.util.Optional;
 @Repository
 public interface BadgeRepository extends MongoRepository<Badge, String> {
 
- Optional<Badge> findByBadgeId(String badgeId);
-    
+    Optional<Badge> findByBadgeId(String badgeId);
+
     List<Badge> findByIsActive(boolean isActive);
 
     // ✅ 必须有这个方法，Service 才能查出同类的其他徽章
@@ -23,5 +24,6 @@ public interface BadgeRepository extends MongoRepository<Badge, String> {
     List<Badge> findByIsActiveAndAcquisitionMethod(boolean isActive, String acquisitionMethod);
 
     // 查找所有启用的、有 carbonThreshold 且阈值 <= 用户 totalCarbon 的徽章
-    List<Badge> findByIsActiveTrueAndAcquisitionMethodAndCarbonThresholdLessThanEqual(String acquisitionMethod, long carbonThreshold);
+    List<Badge> findByIsActiveTrueAndAcquisitionMethodAndCarbonThresholdLessThanEqual(String acquisitionMethod,
+            double carbonThreshold);
 }
